% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_segmenting.R
\name{add_asset_moniker}
\alias{add_asset_moniker}
\title{Add product moniker}
\usage{
add_asset_moniker(assets, filter = NULL)
}
\arguments{
\item{assets}{The assets dataframe to be updated. Requires a 
\code{product_group_detail} column}

\item{filter}{If not the default of NULL, it can contain one or many 
of the following: 'emerging', 'middleware', 'platform', 'other'}
}
\value{
an assets data frame with the new column of product_moniker
}
\description{
Marks assets according to \url{https://pnt.redhat.com/pnt/p-557504/brandarch_2pager_0916EE.pdf}.
Can be then used to segment assets by moniker, or to mark accounts as a
type.
}

