% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_segmenting.R
\name{add_asset_type}
\alias{add_asset_type}
\title{Add asset type}
\usage{
add_asset_type(assets, filter = NULL)
}
\arguments{
\item{assets}{The assets dataframe to be updated. Requires a 
\code{description} and a \code{product_support_type} column}

\item{filter}{If not the default of NULL, it can contain one or many 
of the following: 'free', 'academic', 'L3', 'employee', 'self-support',
'dev suite', 'partner', 'customer'}
}
\value{
an assets data frame with the new columns of asset_type and asset_type_num
}
\description{
Marks assets according to \url{https://mojo.redhat.com/docs/DOC-1005041}.
Can be then used to filter out assets by type, or to mark accounts as a
type.
}

