% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters.R
\name{current_redhat_qtr}
\alias{current_redhat_qtr}
\title{Get the current red hat quarter for today}
\usage{
current_redhat_qtr()
}
\description{
Get the current red hat quarter for today
}
\examples{
x <- current_redhat_qtr()
}

