% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cee_connect.R
\name{dfs_to_csv}
\alias{dfs_to_csv}
\title{Write a bunch of data frames to csv}
\usage{
dfs_to_csv(outdir, pattern)
}
\arguments{
\item{outdir}{the directory to drop the dataframes}

\item{pattern}{a pattern for dataframes to output. The default is to
include all data with the suffix ".df". Works well with \code{\link{run_scripts}}.}
}
\value{
nothing is returned. The output will be csv files.
}
\description{
Write data frames matching a pattern to separate csv files.
}
\details{
Write data frames matching a pattern to separate csv files. Will
write all matching the pattern of '.df' by default.
}

