% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters.R
\name{previous_redhat_qtr}
\alias{previous_redhat_qtr}
\title{Get the red hat quarter previous to the current quarter}
\usage{
previous_redhat_qtr()
}
\description{
Get the red hat quarter previous to the current quarter
}
\examples{
x <- previous_redhat_qtr()
}

