% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cee_connect.R
\name{read_scripts}
\alias{read_scripts}
\title{Read SQL scripts from a directory}
\usage{
read_scripts(base_path, sqlvars, prefix, pattern)
}
\arguments{
\item{base_path}{the directory holding the sql files}

\item{sqlvars}{a list containing all variables you wish to set within your queries}

\item{prefix}{a prefix from your SQL filenames if you want to remove before
creating the data frames}

\item{pattern}{a suffix to include in the data frame names}
}
\value{
sql.list returns a list of SQL queries with the future filename as the name
}
\description{
Read SQL scripts from a directory into R for use against a database.
}
\details{
Just place all scripts in the same directory with a .sql extension, and provide
the directory path. It's also possible to set any variables in your query
with the format of ${variable}.
}

