% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recent_file}
\alias{recent_file}
\title{Determine the most recently saved file}
\usage{
recent_file(directory, reg_ex)
}
\arguments{
\item{directory}{the location of the files in question.}

\item{reg_ex}{the regular expression to search for in the directory}
}
\value{
recent_file returns a filename (including path) that can be then read into R
}
\description{
With many similar filenames, find the most recently saved version
}
\details{
This function asks for a directory and a wildcard regular expression in order to
search a directory for the most recently saved file. This is helpful if you want
to keep historical versions of csv files or data files in the same place, but
want to use the most recent in your calculations.

\code{vdm.conn <- redhat_db_conn("~/Documents/R_code/", "output")}
}

