% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{redhat_colors}
\alias{redhat_colors}
\title{Collect hexidecimal colors}
\usage{
redhat_colors(color)
}
\arguments{
\item{color}{The Red Hat color(s) to be converted. Either a single color
name, or a vector of color names.}
}
\description{
Collect hexidecimal colors
}
\examples{
# Get one color
x <- redhat_colors("Purple")

# Get a vector of colors
y <- redhat_colors(c("Storage 2", "Gray 3", "Red Hat Red", "Gray 10"))
}

