% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cee_connect.R
\name{redhat_db_conn}
\alias{redhat_db_conn}
\title{Connect to a database in R}
\usage{
redhat_db_conn(filename)
}
\arguments{
\item{filename}{the location of the database config file. Should contain the scheme,
host, port, database (if applicable), classname, jars, and user. Password will be
prompted.}
}
\description{
Connect to a database in R
}
\details{
This function requires the user to have access to the database, have downloaded
the driver, and set up the certificates on their computer. The process for using
the package is to create a config file, then pass it along with the path:

\code{vdm.conn <- redhat_db_conn("~/vdm.cfg").}

Then, the created connection can be used with the DBI package:

\code{DBI::dbGetQuery(vdm.conn, "SELECT * FROM APL_CEE_VDM.stg_gss_case LIMIT 10")}
}

