% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters.R
\name{redhat_qtr}
\alias{redhat_qtr}
\title{Old function to get the Red Hat dates}
\usage{
redhat_qtr(x = Sys.Date())
}
\arguments{
\item{x}{The date-type value to be converted. Either a date, date-type
string like '2016-04-25', or a yearmon using the zoo::as.yearmon()
function.}
}
\description{
Old function to get the Red Hat dates
}

