% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters.R
\name{redhatqtr}
\alias{redhatqtr}
\title{Get a Red Hat quarter for a given date or vector of dates}
\usage{
redhatqtr(x = Sys.Date(), previous = FALSE, old = FALSE)
}
\arguments{
\item{x}{The date-type value to be converted. Either a date, date-type
string like '2016-04-25', or a yearmon using the zoo::as.yearmon()
function.}

\item{previous}{A logical scalar. Do you want the previous Red Hat quarter?}

\item{old}{A logical scalar. Do you want the old Q\%qFY\%y format?}
}
\value{
Returns a Red Hat quarter or vector or quarters.
If \code{old = TRUE}, then returns the Red Hat quarter in the old Q\%qFY\%y format.
}
\description{
Get a Red Hat quarter for a given date or vector of dates
}
\details{
The function can take muliple date types (see examples), and returns a Red Hat
quarter.
}
\examples{
redhatqtr()
redhatqtr("2016-04-25")
y <- as.Date("2016-04-12")
redhatqtr(y)
z <- zoo::as.yearmon("2016-01-01")
redhatqtr(z)
redhatqtr(previous = TRUE)
redhatqtr(old = TRUE)
}

