% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters.R
\name{redhatqtr_start}
\alias{redhatqtr_start}
\title{Get the first date of a quarter based on a Red Hat quarter}
\usage{
redhatqtr_start(x = redhatqtr(Sys.Date()))
}
\arguments{
\item{x}{The red hat quarter to be converted. Either the old format like 'Q1FY17',
or the new format like 'FY17Q1'.}
}
\description{
Get the first date of a quarter based on a Red Hat quarter
}
\examples{
redhatqtr_start()
redhatqtr_start("FY17Q1")
redhatqtr_start("Q1FY17")
}

