% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{rh_qtr_year_start}
\alias{rh_qtr_year_start}
\title{Get get the start date of a given Red Hat fiscal quarter.}
\usage{
rh_qtr_year_start(x)
}
\arguments{
\item{x}{The quarter for which to calculate the end of the fiscal year.}
}
\description{
Get get the start date of a given Red Hat fiscal quarter.
}
\examples{
x <- rh_qtr_year_start("Q1FY18")
y <- rh_qtr_year_start(previous_redhat_qtr())
}

