% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cee_connect.R
\name{run_scripts}
\alias{run_scripts}
\title{Run SQL scripts on a database}
\usage{
run_scripts(conn, sql.list)
}
\arguments{
\item{conn}{the database connection created using \code{\link{redhat_db_conn}}}

\item{sql.list}{a list of queries created using \code{\link{read_scripts}}}
}
\value{
nothing is returned. The output from the scripts will be added to your environment.
}
\description{
Run SQL scripts that are stored in a list on a database.
}
\details{
After running \code{\link{redhat_db_conn}} and \code{\link{read_scripts}},
use these parameters to read the database query data into R as data frames.
}

