% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cee_connect.R
\name{set_env_pass}
\alias{set_env_pass}
\title{Create an obscured password to keep in as an environment variable}
\usage{
set_env_pass(passname)
}
\arguments{
\item{passname}{the name of the password (and therefore environment variable) to
be stored as an R environment variable.}
}
\description{
This function allows you to enter your password, and then stores a slightly
obscured version as an environment variable
}
\details{
Since passwords are annoying to enter, and you also don't want them stored as-is,
even in an environment variable file, you can at least mask it a little bit. These
will be saved in a place that only R uses, so if you use python you'll have to do
that separately.
}

