% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show_redhat_colors}
\alias{show_redhat_colors}
\title{Show available Red Hat colors.}
\usage{
show_redhat_colors()
}
\description{
Show available Red Hat colors.
}
\examples{
# Show all Red Hat colors
show_redhat_colors()

}

