% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{slash_paste}
\alias{slash_paste}
\title{Combine portions of directories or urls with one slash}
\usage{
slash_paste(...)
}
\arguments{
\item{...}{strings to be pasted with one slash in between.}
}
\value{
slash_paste returns a string with only one slash between parts
}
\description{
Sometimes you can have a trailing slash, or not enough,
}

