% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{year_end_date}
\alias{year_end_date}
\title{Get get the end date of the current Red Hat fiscal year, or of a specific date.}
\usage{
year_end_date(x)
}
\arguments{
\item{x}{The date for which to calculate the end of the fiscal year.}
}
\description{
Get get the end date of the current Red Hat fiscal year, or of a specific date.
}
\examples{
x <- year_end_date("2016-04-25")
y <- year_end_date(Sys.Date())
}

