% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{year_start_date}
\alias{year_start_date}
\title{Get get the start date of the current Red Hat fiscal year, or of a specific date.}
\usage{
year_start_date(x)
}
\arguments{
\item{x}{The date for which to calculate the beginning of the fiscal year.}
}
\description{
Get get the start date of the current Red Hat fiscal year, or of a specific date.
}
\examples{
x <- year_start_date("2016-04-25")
y <- year_start_date(Sys.Date())
}

