% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters.R
\name{redhat_qtr}
\alias{redhat_qtr}
\title{Get the red hat quarter for a given date, "as.yearmon", or a date-like string}
\usage{
redhat_qtr(x)
}
\arguments{
\item{x}{The date-type value to be converted. Either a date, date-type
string like '2016-04-25', or a yearmon using the zoo::as.yearmon()
function.}
}
\description{
Get the red hat quarter for a given date, "as.yearmon", or a date-like string
}
\examples{
x <- redhat_qtr("2016-04-25")
y <- as.Date("2016-04-12")
y <- redhat_qtr(y)
z <- zoo::as.yearmon("2016-01-01")
z <- redhat_qtr(z)
}

